+++ webapp/cache.py
@@ -1,0 +1,3 @@
+from flask_caching import Cache
+
+cache = Cache()
