+++ webapp/templates/marketplace/add_product.html
@@ -4,9 +4,9 @@
     {% with messages = get_flashed_messages() %}
         {% if messages %}
             {% for message in messages %}
-            <div class="alert alert-danger" role="alert">    
-                {{ message }}<br>
-            </div>
+                <div class="alert alert-danger" role="alert">
+                    {{ message }}<br>
+                </div>
             {% endfor %}
         {% endif %}
     {% endwith %}
